attempt to fix overflow error in get()
The average calcualted as`SUM(values) / NUM(values)` can lead to an
overflow error. This patch divides the values by the number of values
first, then sums them up. Potentially this sacrifices some precision for
the benefit of avoiding an overflow error in most (but not all) cases.

This problem is particularly visible for smaller data types such
as `int16_t`, so possibly the issue #5 is related.